Refactor ExpressionTypes in daft.execution.operators to daft.types
