+++ daft/runners/partitioning.py
@@ -22,11 +22,11 @@ import pandas as pd
 import pyarrow as pa
 from pyarrow import dataset as pada
 
+from daft.execution.operators import OperatorEnum
 from daft.expressions import ColID, ColumnExpression, Expression, ExpressionExecutor
 from daft.logical.schema import ExpressionList
 from daft.runners.blocks import ArrowArrType, DataBlock, PyListDataBlock
-
-from ..execution.operators import OperatorEnum, PythonExpressionType
+from daft.types import ExpressionType
 
 PartID = int
 
@@ -159,11 +159,7 @@ class vPartition:
         for col_expr in column_exprs:
             col_id = col_expr.get_id()
             col_name = col_expr.name()
-            arr = (
-                data[col_name]
-                if isinstance(col_expr.resolved_type(), PythonExpressionType)
-                else pa.array(data[col_expr.name()])
-            )
+            arr = data[col_name] if ExpressionType.is_py(col_expr.resolved_type()) else pa.array(data[col_expr.name()])
             block: DataBlock = DataBlock.make_block(arr)
             tiles[col_id] = PyListTile(column_id=col_id, column_name=col_name, partition_id=partition_id, block=block)
         return vPartition(columns=tiles, partition_id=partition_id)
